General functions have been moved out from the `engine.h`.
  * Macros and some utility functions have been moved out
    from the `engine.h` file to 'core.h'.
