+++ include/engine.h
@@ -1,16 +1,7 @@
 #ifndef ENGINE_H
 #define ENGINE_H
 
-#define ENGINE_USE_INLINE
-#ifdef ENGINE_USE_INLINE
-#ifdef __GNUC__
-#   define ENGINE_INLINE inline __attribute__((always_inline))
-#else
-#   define ENGINE_INLINE inline
-#endif // __GNUC__
-#else
-#   define ENGINE_INLINE
-#endif // ENGINE_USE_INLINE
+#include <core/core.h>
 
 ///
 /// Backing Simulator
@@ -42,13 +33,5 @@ typedef simtime_t timestamp_t;
  * @param event      the event
  */
 ENGINE_INLINE void schedule_event(sid_t service, timestamp_t time, unsigned event_type, void *event);
-
-/**
- * @brief It sends a formatted message to the standard
- *        error and aborts the program.
- *
- * @param fmt the message
- */
-void die(const char *fmt, ...);
 
 #endif // ENGINE_H
